requirements.txt
Flask==2.3.2
Werkzeug==3.1.3
typing-extensions==4.14.1
zipp==3.23.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
